pyproject.toml
[build-system]
requires = ["setuptools>=64.0"]
build-backend = "setuptools.build_meta"

[project]
name = "deepdrivewe"
version = "0.1.0"
authors = [
    {name = "Alexander Brace", email = "abrace@anl.gov"},
    {name = "Nicolas Frazee", email = "ncf26@pitt.edu"},
    {name = "Jeremy Leung", email = "jml230@pitt.edu"},
]
description = "Implementation of DeepDriveWE using Academy."
readme = "README.md"
requires-python = ">=3.8"
license = {text = "MIT"}
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
]
dependencies = [
    "parsl>=2024.10.14",
    "pyyaml>=6.0.1",
    "typer>=0.12.5",
    "numpy",
    "h5py==3.14.0",
    "mdtraj==1.10.3",
    "MDAnalysis>=2.7.0",
    "ParmEd>=4.3.1",
    "scikit-learn==1.6.1",
    "mdlearn==1.0.5",
    "scipy==1.14.0",
    "natsort>=8.4.0",
    "matplotlib>=3.9.2",
    "academy-py>=0.1.0",
]

[project.urls]
homepage = "https://github.com/ramanathanlab/deepdrivewe-academy"
documentation = "https://github.com/ramanathanlab/deepdrivewe-academy"
repository = "https://github.com/ramanathanlab/deepdrivewe-academy"

[project.optional-dependencies]
dev = [
    "covdefaults>=2.2",
    "coverage",
    "mypy",
    "pre-commit",
    "pytest",
    "pytest-asyncio",
    "pytest-cov",
    "ruff",
    "tox",
    "virtualenv",
    "types-PyYAML",
]
docs = [
    "black",
    "properdocs",
    "mkdocs-gen-files",
    "mkdocs-literate-nav",
    "mkdocs-material>=9.5",
    "mkdocs-section-index",
    "mkdocstrings>=0.26",
    "mkdocstrings-python>=1.12",
    "mike",
]

[tool.codespell]
skip = """
.git,
.github,
__pycache__,
build,
dist,
.*egg-info
"""

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
markers = [
    "unit: fast, hermetic tests with no external I/O",
    "integration: tests that hit the filesystem or HDF5",
]

[tool.coverage.run]
plugins = ["covdefaults"]
source = ["deepdrivewe"]
omit = [
    "examples/*",
    # Heavy-infrastructure modules not covered by the unit suite. These
    # require OpenMM/Amber/torch and are exercised via example runs, not
    # pytest. Revisit once CI has those deps.
    "deepdrivewe/simulation/*",
    "deepdrivewe/ai/*",
    "deepdrivewe/parsl.py",
]

[tool.coverage.report]
# Coverage target for the core math/policy modules (api, binners,
# resamplers, recyclers, utils, checkpoint, io). Bump toward 80%+ as
# the suite grows.
fail_under = 70
precision = 2

[tool.mypy]
python_version = "3.10"
explicit_package_bases = true
namespace_packages = true
check_untyped_defs = true
disallow_any_generics = true
disallow_incomplete_defs = true
disallow_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_configs = true
warn_unused_ignores = true

[tools.mypy]
plugins = ["academy.mypy_plugin"]

[[tool.mypy.overrides]]
module = "testing.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[[tool.mypy.overrides]]
module = "tests.*"
allow_incomplete_defs = true
allow_untyped_defs = true

[tool.ruff]
# See all rules here: https://beta.ruff.rs/docs/rules
select = [
    # pyflakes
    "F",
    # pycodestyle
    "E",
    # mccabe
    "C90",
    # isort
    "I",
    # pep8-naming
    "N",
    # pydocstyle
    "D",
    # pyupgrade
    "UP",
    # flake8-2020
    "YTT",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-commas
    "COM",
    # flake8-comprehensions
    "C4",
    # flake8-implicit-str-concat
    "ISC",
    # flake8-pytest-style
    "PT",
    # flake8-quotes
    "Q",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # PyLint
    "PL",
    # ruff-specific
    "RUF",
]
line-length = 79
extend-ignore = ["Q001"]
target-version = "py310"

[tool.ruff.flake8-pytest-style]
parametrize-values-type = "tuple"

[tool.ruff.flake8-quotes]
inline-quotes = "single"
multiline-quotes = "single"

[tool.ruff.format]
indent-style = "space"
quote-style = "single"

[tool.ruff.isort]
force-single-line = true
known-first-party = ["deepdrivewe", "test", "testing", "examples"]
order-by-type = false
required-imports = ["from __future__ import annotations"]

[tool.ruff.per-file-ignores]
"*/__init__.py" = ["F401"]
# Tests are allowed to use magic numbers (PLR2004), broad pytest.raises
# (PT011), broad pytest.warns (PT030), and more relaxed docstring
# conventions (D10, D401).
"*/test_*.py" = ["D10", "D401", "PLR2004", "PT011", "PT030", "SLF001"]
"tests/conftest.py" = ["D401"]
# Lazy imports inside functions are intentional in these modules:
# heavy deps (torch, mdlearn, synd) and optional plugins
# (globus-compute-sdk) are imported on-demand to keep the local
# install lightweight.
"deepdrivewe/ai/*.py" = ["PLC0415"]
"deepdrivewe/simulation/synd.py" = ["PLC0415"]
"examples/**/*.py" = ["PLC0415"]

[tool.ruff.lint.pylint]
max-args = 8

[tool.ruff.pydocstyle]
convention = "numpy"

[tool.setuptools.packages.find]
exclude = ["tests*", "testing*"]
namespaces = false
